move src/url_util.ts to src/lib/
